Fix FileUtils.fileContentAsString() Fix PrefixArray.init<>()

+++ src/main/java/yapi/datastructures/PrefixArray.java
@@ -1,6 +1,5 @@
 package yapi.datastructures;
 
-import com.sun.imageio.plugins.common.I18N;
 import yapi.exceptions.ArrayMutationException;
 import yapi.math.NumberUtils;
 
@@ -28,7 +27,7 @@ public class PrefixArray<E> {
     }
 
     public PrefixArray(String s) {
-        List<Integer> range = NumberUtils.getRange(s);
+        List<Long> range = NumberUtils.getRange(s);
         for (int i = 0; i < range.size(); i++) {
             if (i == 0) {
                 head = (E)range.get(i);
